+++ runtime/src/test/scala/com/ing/baker/runtime/actor/ProcessInstanceSpec.scala
@@ -4,14 +4,12 @@ import java.util.UUID
 import java.util.concurrent.TimeUnit
 
 import akka.actor.{ActorRef, ActorSystem, PoisonPill, Terminated}
-import akka.pattern.ask
 import akka.testkit.TestDuration
 import akka.util.Timeout
 import com.ing.baker.petrinet.api._
 import com.ing.baker.petrinet.dsl.colored._
 import com.ing.baker.petrinet.runtime.ExceptionStrategy.{BlockTransition, Fatal, RetryWithDelay}
 import com.ing.baker.petrinet.runtime.TransitionExceptionHandler
-import com.ing.baker.runtime.actor.AkkaTestBase.GetChild
 import com.ing.baker.runtime.actor.ProcessInstanceProtocol._
 import org.mockito.Matchers._
 import org.mockito.Mockito._
@@ -94,6 +92,7 @@ class ProcessInstanceSpec extends AkkaTestBase with ScalaFutures with MockitoSug
       watch(actor)
       actor ! GetState
       expectMsgClass(classOf[Uninitialized])
+      expectMsgClass(classOf[Terminated])
     }
 
     "After being initialized respond with an InstanceState message on receiving a GetState command" in new TestSequenceNet {
@@ -392,18 +391,14 @@ class ProcessInstanceSpec extends AkkaTestBase with ScalaFutures with MockitoSug
       )
 
       val petriNetActor = createPetriNetActor(coloredProps(petriNet, runtime, customSettings), UUID.randomUUID().toString)
+      watch(petriNetActor)
 
       implicit val timeout = Timeout(dilatedMillis(2000), MILLISECONDS)
-      whenReady((petriNetActor ? GetChild).mapTo[ActorRef]) {
-        child ⇒
-          {
-            petriNetActor ! Initialize(marshal[Place](initialMarking), ())
-            expectMsgClass(classOf[Initialized])
 
-            watch(child)
-            expectMsgClass(classOf[Terminated])
-          }
-      }
+      petriNetActor ! Initialize(marshal[Place](initialMarking), ())
+      expectMsgClass(classOf[Initialized])
+
+      expectMsgClass(classOf[Terminated])
     }
 
     "fire automated transitions in parallel when possible" in new StateTransitionNet[Unit, Unit] {