+++ CHANGELOG.md
@@ -1,12 +1,28 @@
 # Changelog
 
+## 1.1.15
+- Fix #62: internal IdleStop message for the ProcessInstance actor is now configured to be serialized by Kryo
+- Fix #59: disabled the usage of Akka Distributed Data until the growing memory issue in the shared process metadata feature is solved.
+- Fix #57: configured the default actor idle-timeout as 5 minutes (baker.actor.idle-timeout config can be overridden in the application.conf)
+- Fix #56: fixed one unhandled message warning for ProcessInstanceEvent
+- deprecated the methods that return the generated SVG
+- cleaned up unnecessary Passivate pattern from ProcessInstance
+- removed unnecessary debug logging of Akka Distributed Data updates
+- log when scheduling a retry on actor startup
+
+## 1.1.14
+- Fix #53: EventListeners are now notified of retry-exhausted events.
+- Fix #49: improved error message when receiving invalid sensory event
+- Added a method to CompiledRecipe to obtain an SVG String: ```getVisualRecipeAsSVG```
+- Updated to akka 2.5.6
+
 ## 1.1.13
-- Fix for issue #47: added writeVisualrecipeToSVGFile to write away the CompiledRecipe to a file.
-- bugfix for issue #46: it's now allowed to require an event on the name only, however it's still possible to require on event class in the JavaDSL.
-- Fix for issue #45: validate that ingredients are not of primitive type after compilation.
+- Fix #47: added writeVisualrecipeToSVGFile to write away the CompiledRecipe to a file.
+- Fix #46: it's now allowed to require an event on the name only, however it's still possible to require on event class in the JavaDSL.
+- Fix #45: validate that ingredients are not of primitive type after compilation.
 
 ## 1.1.12
-- bugfix for issue #43: interaction is not compiled into petrinet when requires a renamed optional ingredient provided through a renamed event
+- Fix #43: interaction is not compiled into petrinet when requires a renamed optional ingredient provided through a renamed event
 
 ## 1.1.11
 - bugfix: predefining a higher order type other then option/optional crashed the recipe validations when compiling a recipe